Get ring detection and contour center finder up and running on Control Hub
Ring detector class was created (just old WebcamTest class renamed), which will detect rings and draw a bounding box around their contours.
It will also find and draw a square at the center of the bounding box, which will be used to guide the robot. I fixed the memory leak issues that were causing the stream to not update and fail, so now the pipeline's mat output can be seen by hooking up a display to the Control Hub or by selecting the Camera Stream option after pressing init in the Driver Station app.